+++ tests/rules/named-export.test.ts
@@ -1,3 +1,5 @@
+import path from 'path';
+
 import { RuleTester } from 'eslint';
 import { namedExport } from '#/rules/named-export';
 
@@ -77,6 +79,53 @@ ruleTesterES2015.run('named-export: single named export', namedExport, {
       code: 'export const rule = {}',
       filename: 'src/rules/index.js',
       errors: ['The export name must match the filename. You need to rename to Rules or rules.'],
+    },
+  ],
+});
+
+const ruleTesterTS2015 = new RuleTester({
+  parser: path.resolve(__dirname, '../../node_modules/@typescript-eslint/parser'),
+  parserOptions: { ecmaVersion: '2015', sourceType: 'module' },
+});
+
+ruleTesterTS2015.run('named-export(ts): default export', namedExport, {
+  valid: [
+    {
+      code: 'const module: number = 1; export default module;',
+      filename: 'module.ts',
+    },
+  ],
+  invalid: [],
+});
+
+ruleTesterTS2015.run('named-export(ts): multiple named export', namedExport, {
+  valid: [
+    {
+      code: "export const module1: number = 1; export const module2: string = 'string'",
+      filename: 'module.ts',
+    },
+    {
+      code: `const module1: any[] = [];
+const module2: { key: string } = { key: 'string' };
+export { module1, module2 }`,
+      filename: 'module.ts',
+    },
+  ],
+  invalid: [],
+});
+
+ruleTesterTS2015.run('named-export(ts): single named export', namedExport, {
+  valid: [
+    {
+      code: 'export type Empty = {}',
+      filename: 'empty.ts',
+    },
+  ],
+  invalid: [
+    {
+      code: 'export type Blank = {}',
+      filename: 'empty.ts',
+      errors: ['The export name must match the filename. You need to rename to Empty or empty.'],
     },
   ],
 });