+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Unreleased
+## Bugfix
+* [#27](https://github.com/epaew/eslint-plugin-filenames-simple/pull/27) Fixed the behavior that ESLint execution fails
+when the lint target includes TypeScript notation and the rule `named-export` is enabled.
 
 # 0.3.0
 ## Features